adding google and facebook login , style fixing

+++ crime/views.py
@@ -2,6 +2,8 @@ from __future__ import division
 from django.shortcuts import HttpResponse, render, HttpResponseRedirect
 from django.template import RequestContext
 from django.db.models.query import Q
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 from crime.forms import crimeForm
 from crime.models import crime
 from geoposition import Geoposition
@@ -11,7 +13,7 @@ from calendar import timegm
 
 
 def acceuil(request):
-    context_dict = {'hello': "Introduction"}
+    context_dict = {'hello': "Introduction", 'p': 'p1'}
     return render(request, 'crime/acceuil.html', context_dict)
 
 
@@ -20,6 +22,7 @@ def thanks(request):
     return render(request, 'crime/thanks.html', context_dict)
 
 
+@login_required
 def add_crime(request):
     context = RequestContext(request)
     if request.method == 'POST':
@@ -27,18 +30,22 @@ def add_crime(request):
         var = request.POST['crimetype']
 
         if form.is_valid():
-            form.save(commit=True)
+            form = form.save(commit=False)
+            form.personne = request.user
+            form.save()
             return thanks(request)
         else:
             print form.errors
     else:
 
         form = crimeForm()
+    p = 'p2'
 
     return render(request, 'crime/add_crime.html', locals())
 
 
 def statistique(request):
+    p = 'p5'
     return render(request, 'crime/statistique.html', locals())
 
 
@@ -48,7 +55,8 @@ def delete(request):
     the_id = int(request.GET['id'])
     try:
         obj = crime.objects.get(id=the_id)
-        obj.delete()
+        if obj.personne == request.user:
+            obj.delete()
         return HttpResponse("success ")
     except crime.DoesNotExist:
         return HttpResponse("fail")
@@ -56,13 +64,14 @@ def delete(request):
 
 def modify(request):
     if request.method == 'GET':
-        if 'lat' and 'lng' not in request.GET:
+        if 'lat' and 'lng' and 'uid' not in request.GET:
             return HttpResponse("fail no position in get")
         lat = request.GET['lat']
         lng = request.GET['lng']
+        uid = request.GET['uid']
         try:
             p = Geoposition(lat, lng)
-            obj = crime.objects.get(position=p)
+            obj = crime.objects.get(position=p, personne=User.objects.get(id=uid))
             obj_id = obj.id
             gouv = obj.gouvernorat
             desc = obj.description
@@ -92,19 +101,20 @@ def modify(request):
             obj = crime.objects.get(id=the_id)
         except crime.DoesNotExist:
             return HttpResponse("fail no crime")
-        if "gouvernorat" in request.POST:
-            obj.gouvernorat = request.POST["gouvernorat"]
-        if "description" in request.POST:
-            obj.description = request.POST["description"]
-        if "sexe" in request.POST:
-            obj.sexe = request.POST["sexe"]
-        if "crimetype" in request.POST:
-            obj.crimetype = request.POST["crimetype"]
-        if "position" in request.POST:
-            position = request.POST["position"]
-        if "time" in request.POST:
-            obj.time = request.POST["time"]
-        obj.save()
+        if request.user == obj.personne:
+            if "gouvernorat" in request.POST:
+                obj.gouvernorat = request.POST["gouvernorat"]
+            if "description" in request.POST:
+                obj.description = request.POST["description"]
+            if "sexe" in request.POST:
+                obj.sexe = request.POST["sexe"]
+            if "crimetype" in request.POST:
+                obj.crimetype = request.POST["crimetype"]
+            if "position" in request.POST:
+                position = request.POST["position"]
+            if "time" in request.POST:
+                obj.time = request.POST["time"]
+            obj.save()
         return HttpResponseRedirect('/crime/mod/')
     else:
         return HttpResponse("fail all")
@@ -112,13 +122,15 @@ def modify(request):
 
 def filtercrime(request):
     if request.method == "GET":
-        sex, gouv, cat = None, None, None
+        sex, gouv, cat, uid = None, None, None, None
         if "sex" in request.GET:
             sex = request.GET["sexe"]
         if "gouvernorat" in request.GET:
             gouv = request.GET["gouvernorat"]
         if "crimetype" in request.GET:
             cat = request.GET["crimetype"]
+        if "uid" in request.GET:
+            uid = request.GET["uid"]
         q = Q()
         if sex:
             q = q & Q(sexe=sex)
@@ -126,6 +138,8 @@ def filtercrime(request):
             q = q & Q(gouvernorat=gouv)
         if cat:
             q = q & Q(crimetype=cat)
+        if uid:
+            q = q & Q(personne=User.objects.get(id=uid))
         crimes = crime.objects.filter(q)
 
         crimes = list(crimes.values_list("gouvernorat", "position", "crimetype", "time"))
@@ -169,12 +183,16 @@ def calendar_data(request):
 
 
 def cal(request):
+    p = 'p6'
     return render(request, 'crime/calendar.html', locals())
 
 
 def mod_crime(request):
+    p = 'p3'
     form = crimeForm()
     return render(request, 'crime/mod_crime.html', locals())
 
+
 def carte(request):
+    p = 'p4'
     return render(request, 'crime/carte.html', locals())
